Rewrite storageKey() & refactoring
